+++ examples/02-inputs/02-decoding-events/tailwind.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  content: ["./index.html", "./src/**/*.{gleam,mjs}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};