+++ CHANGELOG.md
@@ -14,6 +14,7 @@ request to fix it.
 ### Changed
 
 - [lustre/element] `element.advanced` correctly respects the `self_closing` and `void` options when generating HTML strings.
+- [lustre/event] Fixed a bug where debounced events inside components decoded using the wrong event data.
 
 ## [v5.3.5] - 2025-09-03